game create game into db and move seeding into a new seed file
